+++ unliftio/src/UnliftIO/Async.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE RankNTypes #-}
--- | Unlifted "Control.Concurrent.Async"
+-- | Unlifted "Control.Concurrent.Async".
 --
 -- @since 0.1.0.0
 module UnliftIO.Async
@@ -62,57 +62,57 @@ import Data.Foldable (Foldable)
 import Data.Traversable (Traversable)
 #endif
 
--- | Unlift 'A.async'
+-- | Unlifted 'A.async'.
 --
 -- @since 0.1.0.0
 async :: MonadUnliftIO m => m a -> m (Async a)
 async m = withRunInIO $ \run -> A.async $ run m
 
--- | Unlift 'A.asyncBound'
+-- | Unlifted 'A.asyncBound'.
 --
 -- @since 0.1.0.0
 asyncBound :: MonadUnliftIO m => m a -> m (Async a)
 asyncBound m = withRunInIO $ \run -> A.asyncBound $ run m
 
--- | Unlift 'A.asyncOn'
+-- | Unlifted 'A.asyncOn'.
 --
 -- @since 0.1.0.0
 asyncOn :: MonadUnliftIO m => Int -> m a -> m (Async a)
 asyncOn i m = withRunInIO $ \run -> A.asyncOn i $ run m
 
--- | Unlift 'A.asyncWithUnmask'
+-- | Unlifted 'A.asyncWithUnmask'.
 --
 -- @since 0.1.0.0
 asyncWithUnmask :: MonadUnliftIO m => ((forall b. m b -> m b) -> m a) -> m (Async a)
 asyncWithUnmask m =
   withUnliftIO $ \u -> A.asyncWithUnmask $ \unmask -> unliftIO u $ m $ liftIO . unmask . unliftIO u
 
--- | Unlift 'A.asyncOnWithUnmask'
+-- | Unlifted 'A.asyncOnWithUnmask'.
 --
 -- @since 0.1.0.0
 asyncOnWithUnmask :: MonadUnliftIO m => Int -> ((forall b. m b -> m b) -> m a) -> m (Async a)
 asyncOnWithUnmask i m =
   withUnliftIO $ \u -> A.asyncOnWithUnmask i $ \unmask -> unliftIO u $ m $ liftIO . unmask . unliftIO u
 
--- | Unlift 'A.withAsync'
+-- | Unlifted 'A.withAsync'.
 --
 -- @since 0.1.0.0
 withAsync :: MonadUnliftIO m => m a -> (Async a -> m b) -> m b
 withAsync a b = withUnliftIO $ \u -> A.withAsync (unliftIO u a) (unliftIO u . b)
 
--- | Unlift 'A.withAsyncBound'
+-- | Unlifted 'A.withAsyncBound'.
 --
 -- @since 0.1.0.0
 withAsyncBound :: MonadUnliftIO m => m a -> (Async a -> m b) -> m b
 withAsyncBound a b = withUnliftIO $ \u -> A.withAsyncBound (unliftIO u a) (unliftIO u . b)
 
--- | Unlift 'A.withAsyncOn'
+-- | Unlifted 'A.withAsyncOn'.
 --
 -- @since 0.1.0.0
 withAsyncOn :: MonadUnliftIO m => Int -> m a -> (Async a -> m b) -> m b
 withAsyncOn i a b = withUnliftIO $ \u -> A.withAsyncOn i (unliftIO u a) (unliftIO u . b)
 
--- | Unlift 'A.withAsyncWithUnmask'
+-- | Unlifted 'A.withAsyncWithUnmask'.
 --
 -- @since 0.1.0.0
 withAsyncWithUnmask
@@ -125,7 +125,7 @@ withAsyncWithUnmask a b =
     (\unmask -> unliftIO u $ a $ liftIO . unmask . unliftIO u)
     (unliftIO u . b)
 
--- | Unlift 'A.withAsyncOnWithMask'
+-- | Unlifted 'A.withAsyncOnWithMask'.
 --
 -- @since 0.1.0.0
 withAsyncOnWithUnmask
@@ -139,31 +139,31 @@ withAsyncOnWithUnmask i a b =
     (\unmask -> unliftIO u $ a $ liftIO . unmask . unliftIO u)
     (unliftIO u . b)
 
--- | Lifted 'A.wait'
+-- | Lifted 'A.wait'.
 --
 -- @since 0.1.0.0
 wait :: MonadIO m => Async a -> m a
 wait = liftIO . A.wait
 
--- | Lifted 'A.poll'
+-- | Lifted 'A.poll'.
 --
 -- @since 0.1.0.0
 poll :: MonadIO m => Async a -> m (Maybe (Either SomeException a))
 poll = liftIO . A.poll
 
--- | Lifted 'A.waitCatch'
+-- | Lifted 'A.waitCatch'.
 --
 -- @since 0.1.0.0
 waitCatch :: MonadIO m => Async a -> m (Either SomeException a)
 waitCatch = liftIO . A.waitCatch
 
--- | Lifted 'A.cancel'
+-- | Lifted 'A.cancel'.
 --
 -- @since 0.1.0.0
 cancel :: MonadIO m => Async a -> m ()
 cancel = liftIO . A.cancel
 
--- | Lifted 'A.uninterruptibleCancel'
+-- | Lifted 'A.uninterruptibleCancel'.
 --
 -- @since 0.1.0.0
 uninterruptibleCancel :: MonadIO m => Async a -> m ()
@@ -176,139 +176,139 @@ uninterruptibleCancel = liftIO . A.uninterruptibleCancel
 cancelWith :: (Exception e, MonadIO m) => Async a -> e -> m ()
 cancelWith a e = liftIO (A.cancelWith a (E.toAsyncException e))
 
--- | Lifted 'A.waitAny'
+-- | Lifted 'A.waitAny'.
 --
 -- @since 0.1.0.0
 waitAny :: MonadIO m => [Async a] -> m (Async a, a)
 waitAny = liftIO . A.waitAny
 
--- | Lifted 'A.waitAnyCatch'
+-- | Lifted 'A.waitAnyCatch'.
 --
 -- @since 0.1.0.0
 waitAnyCatch :: MonadIO m => [Async a] -> m (Async a, Either SomeException a)
 waitAnyCatch = liftIO . A.waitAnyCatch
 
--- | Lifted 'A.waitAnyCancel'
+-- | Lifted 'A.waitAnyCancel'.
 --
 -- @since 0.1.0.0
 waitAnyCancel :: MonadIO m => [Async a] -> m (Async a, a)
 waitAnyCancel = liftIO . A.waitAnyCancel
 
--- | Lifted 'A.waitAnyCatchCancel'
+-- | Lifted 'A.waitAnyCatchCancel'.
 --
 -- @since 0.1.0.0
 waitAnyCatchCancel :: MonadIO m => [Async a] -> m (Async a, Either SomeException a)
 waitAnyCatchCancel = liftIO . A.waitAnyCatchCancel
 
--- | Lifted 'A.waitEither'
+-- | Lifted 'A.waitEither'.
 --
 -- @since 0.1.0.0
 waitEither :: MonadIO m => Async a -> Async b -> m (Either a b)
 waitEither a b = liftIO (A.waitEither a b)
 
--- | Lifted 'A.waitEitherCatch'
+-- | Lifted 'A.waitEitherCatch'.
 --
 -- @since 0.1.0.0
 waitEitherCatch :: MonadIO m => Async a -> Async b -> m (Either (Either SomeException a) (Either SomeException b))
 waitEitherCatch a b = liftIO (A.waitEitherCatch a b)
 
--- | Lifted 'A.waitEitherCancel'
+-- | Lifted 'A.waitEitherCancel'.
 --
 -- @since 0.1.0.0
 waitEitherCancel :: MonadIO m => Async a -> Async b -> m (Either a b)
 waitEitherCancel a b = liftIO (A.waitEitherCancel a b)
 
--- | Lifted 'A.waitEitherCatchCancel'
+-- | Lifted 'A.waitEitherCatchCancel'.
 --
 -- @since 0.1.0.0
 waitEitherCatchCancel :: MonadIO m => Async a -> Async b -> m (Either (Either SomeException a) (Either SomeException b))
 waitEitherCatchCancel a b = liftIO (A.waitEitherCatchCancel a b)
 
--- | Lifted 'A.waitEither_'
+-- | Lifted 'A.waitEither_'.
 --
 -- @since 0.1.0.0
 waitEither_ :: MonadIO m => Async a -> Async b -> m ()
 waitEither_ a b = liftIO (A.waitEither_ a b)
 
--- | Lifted 'A.waitBoth'
+-- | Lifted 'A.waitBoth'.
 --
 -- @since 0.1.0.0
 waitBoth :: MonadIO m => Async a -> Async b -> m (a, b)
 waitBoth a b = liftIO (A.waitBoth a b)
 
--- | Lifted 'A.link'
+-- | Lifted 'A.link'.
 --
 -- @since 0.1.0.0
 link :: MonadIO m => Async a -> m ()
 link = liftIO . A.link
 
--- | Lifted 'A.link2'
+-- | Lifted 'A.link2'.
 --
 -- @since 0.1.0.0
 link2 :: MonadIO m => Async a -> Async b -> m ()
 link2 a b = liftIO (A.link2 a b)
 
--- | Unlifted 'A.race'
+-- | Unlifted 'A.race'.
 --
 -- @since 0.1.0.0
 race :: MonadUnliftIO m => m a -> m b -> m (Either a b)
 race a b = withUnliftIO $ \u -> A.race (unliftIO u a) (unliftIO u b)
 
--- | Unlifted 'A.race_'
+-- | Unlifted 'A.race_'.
 --
 -- @since 0.1.0.0
 race_ :: MonadUnliftIO m => m a -> m b -> m ()
 race_ a b = withUnliftIO $ \u -> A.race_ (unliftIO u a) (unliftIO u b)
 
--- | Unlifted 'A.concurrently'
+-- | Unlifted 'A.concurrently'.
 --
 -- @since 0.1.0.0
 concurrently :: MonadUnliftIO m => m a -> m b -> m (a, b)
 concurrently a b = withUnliftIO $ \u -> A.concurrently (unliftIO u a) (unliftIO u b)
 
--- | Unlifted 'A.concurrently_'
+-- | Unlifted 'A.concurrently_'.
 --
 -- @since 0.1.0.0
 concurrently_ :: MonadUnliftIO m => m a -> m b -> m ()
 concurrently_ a b = withUnliftIO $ \u -> A.concurrently_ (unliftIO u a) (unliftIO u b)
 
--- | Unlifted 'A.mapConcurrently'
+-- | Unlifted 'A.mapConcurrently'.
 --
 -- @since 0.1.0.0
 mapConcurrently :: MonadUnliftIO m => Traversable t => (a -> m b) -> t a -> m (t b)
 mapConcurrently f t = withRunInIO $ \run -> A.mapConcurrently (run . f) t
 
--- | Unlifted 'A.forConcurrently'
+-- | Unlifted 'A.forConcurrently'.
 --
 -- @since 0.1.0.0
 forConcurrently :: MonadUnliftIO m => Traversable t => t a -> (a -> m b) -> m (t b)
 forConcurrently t f = withRunInIO $ \run -> A.forConcurrently t (run . f)
 
--- | Unlifted 'A.mapConcurrently_'
+-- | Unlifted 'A.mapConcurrently_'.
 --
 -- @since 0.1.0.0
 mapConcurrently_ :: MonadUnliftIO m => Foldable f => (a -> m b) -> f a -> m ()
 mapConcurrently_ f t = withRunInIO $ \run -> A.mapConcurrently_ (run . f) t
 
--- | Unlifted 'A.forConcurrently_'
+-- | Unlifted 'A.forConcurrently_'.
 --
 -- @since 0.1.0.0
 forConcurrently_ :: MonadUnliftIO m => Foldable f => f a -> (a -> m b) -> m ()
 forConcurrently_ t f = withRunInIO $ \run -> A.forConcurrently_ t (run . f)
 
--- | Unlifted 'A.replicateConcurrently'
+-- | Unlifted 'A.replicateConcurrently'.
 --
 -- @since 0.1.0.0
 replicateConcurrently :: MonadUnliftIO m => Int -> m a -> m [a]
 replicateConcurrently i m = withRunInIO $ \run -> A.replicateConcurrently i (run m)
 
--- | Unlifted 'A.replicateConcurrently_'
+-- | Unlifted 'A.replicateConcurrently_'.
 --
 -- @since 0.1.0.0
 replicateConcurrently_ :: MonadUnliftIO m => Int -> m a -> m ()
 replicateConcurrently_ i m = withRunInIO $ \run -> A.replicateConcurrently_ i (run m)
 
--- | Unlifted 'A.Concurrently'
+-- | Unlifted 'A.Concurrently'.
 --
 -- @since 0.1.0.0
 newtype Concurrently m a = Concurrently
@@ -332,7 +332,7 @@ instance MonadUnliftIO m => Alternative (Concurrently m) where
     Concurrently $ liftM (either id id) (race as bs)
 
 #if MIN_VERSION_base(4,9,0)
--- | Only defined by @async@ for @base >= 4.9@
+-- | Only defined by @async@ for @base >= 4.9@.
 --
 -- @since 0.1.0.0
 instance (MonadUnliftIO m, Semigroup a) => Semigroup (Concurrently m a) where