+++ unliftio-core/ChangeLog.md
@@ -1,7 +1,8 @@
 ## 0.1.1.0
 
-* Doc improvements
+* Doc improvements.
+* Inline functions in `Control.Monad.IO.Unlift` module [#4](https://github.com/fpco/unliftio/pull/4).
 
 ## 0.1.0.0
 
-* Initial release
+* Initial release.