+++ unliftio/ChangeLog.md
@@ -1,8 +1,9 @@
 ## 0.1.1.0
 
-* Doc improvements
-* Fix UnliftIO.Chan type signatures [#3](https://github.com/fpco/unliftio/pull/3)
+* Doc improvements.
+* Fix `UnliftIO.Chan` type signatures [#3](https://github.com/fpco/unliftio/pull/3).
+* Add `UnliftIO.Concurrent` module [#5](https://github.com/fpco/unliftio/pull/5).
 
 ## 0.1.0.0
 
-* Initial release
+* Initial release.